+++ .gitignore
@@ -129,3 +129,4 @@ dist
 .yarn/install-state.gz
 .pnp.*
 .idea
+serviceAccountKey.json
