+++ authentication API/firebase.js
