+++ authentication API/server.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const authRoutes = require("./routes/auth");
+const authRoutes = require("./auth");
 
 const app = express();
 app.use(bodyParser.json());
