+++ package.json
@@ -2,7 +2,7 @@
   "name": "cloud-computing-bangkit",
   "version": "1.0.0",
   "description": "Repository Tim Cloud Computing",
-  "main": "index.js",
+  "main": "authentication API/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
